Authorisation added such that only users with admin roles can post, put, patch or delete data in api  and users with vet roles can view data in api. Non users cannot view any data

+++ routes/api.php
@@ -19,23 +19,23 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::get('/owners', 'API\\Owners@index');
-Route::get('/owners/{owner}', 'API\\Owners@show');
+Route::get('/owners', 'API\\Owners@index')->middleware('auth:api');
+Route::get('/owners/{owner}', 'API\\Owners@show')->middleware('auth:api');
 
-Route::delete('/owners/{owner}', 'API\\Owners@destroy');
+Route::delete('/owners/{owner}', 'API\\Owners@destroy')->middleware('auth:api');
 
-Route::post('/owners', 'API\\Owners@store');
+Route::post('/owners', 'API\\Owners@store')->middleware('auth:api');
 
-Route::put('/owners/{owner}', 'API\\Owners@update');
+Route::put('/owners/{owner}', 'API\\Owners@update')->middleware('auth:api');
 
-Route::get('/animals', 'API\\Animals@index');
-Route::get('/animals/{animal}', 'API\\Animals@show');
+Route::get('/animals', 'API\\Animals@index')->middleware('auth:api');
+Route::get('/animals/{animal}', 'API\\Animals@show')->middleware('auth:api');
 
-Route::delete('/animals/{animal}', 'API\\Animals@destroy');
+Route::delete('/animals/{animal}', 'API\\Animals@destroy')->middleware('auth:api');
 
-Route::post('/animals', 'API\\Animals@store');
+Route::post('/animals', 'API\\Animals@store')->middleware('auth:api');
 
-Route::put('/animals/{animal}', 'API\\Animals@update');
+Route::put('/animals/{animal}', 'API\\Animals@update')->middleware('auth:api');
 
-Route::get('/owners/{owner}/animals', 'API\\Owners\\Animals@show');
-Route::post('/owners/{owner}/animals', 'API\\Owners\\Animals@store');
+Route::get('/owners/{owner}/animals', 'API\\Owners\\Animals@show')->middleware('auth:api');
+Route::post('/owners/{owner}/animals', 'API\\Owners\\Animals@store')->middleware('auth:api');
